requirements.txt
pandas
boto3
botocore
requests
azure-identity
azure-mgmt-resource
azure-mgmt-compute
azure-mgmt-network
azure-storage-blob
azure-mgmt-security
google-cloud
google-auth
termcolor
colorama
